requirements.txt
# ── 墨麟OS v3.0 Python依赖清单 ──────────────────────────────
# 安装命令: python3.11 -m pip install -r requirements.txt

# 核心AI/LLM
openai>=1.30.0
anthropic>=0.28.0
dashscope>=1.20.0

# 向量数据库（记忆存储）
qdrant-client>=1.9.0

# 记忆增强
mem0ai>=0.1.0

# 网页抓取（双引擎）
firecrawl-py>=1.0.0
crawl4ai>=0.4.0
playwright>=1.44.0

# 浏览器自动化
browser-use>=0.1.0

# HTTP客户端
httpx>=0.27.0
requests>=2.32.0
aiohttp>=3.9.0

# 数据处理
pandas>=2.2.0
numpy>=1.26.0

# 环境配置
python-dotenv>=1.0.0

# CLI和终端
rich>=13.7.0
typer>=0.12.0
click>=8.1.7

# 文件处理
PyYAML>=6.0.1
toml>=0.10.2
python-frontmatter>=1.1.0

# Markdown处理
markdown>=3.6.0
mistune>=3.0.2

# 日期时间
python-dateutil>=2.9.0

# 异步支持
asyncio-mqtt>=0.16.2

# 图像处理（视觉内容）
Pillow>=10.3.0

# 音频处理
pydub>=0.25.1

# 数据存储
SQLAlchemy>=2.0.30

# 飞书SDK（Python版，可选）
lark-oapi>=1.3.0
